Goto start.spring.io and generate project
In the first step, we go to http://start.spring.io and generate a project containing

* web
* jOOQ
* H2

this will get us so called startes for all those libraries. Those starters provide automatic configuration for their libraries. We then have an embedded container (Tomcat), working Spring MVC and an embedded H2 database, so that this project actually compiles and runs without errors.

I also prefer to have the Spring Boot devtools on the classpath, so that I don't have to restart the app during development.