Configure the jOOQ code generator
Run the jOOQ code generator after sucessfull migration of the database. Use flyways version table as input for the jOOQ version to avoid frequent regeneration of code.

Generate only the DOAG2016 schema but generate the classes to the public schema, so that the resulting application will be able to run against a different named schema.

Also generate immutable pojos.