Create reporting controller for Artists
This is one way to "connect" jOOQ to the web: For this demo i opted for jOOQs #formatJSON method, writing directly to the HttpServletResponse, mainly to keep the demo short.

Most of the time I prefer encapsulating the DSLContext inside some @Repository bean, which not only gives me transparent exception handling but can also be easier mocked away in tests than the DSLContext. I would use the generated POJOs in the repository, to retrieve and insert / update my data.

As you may have notived, i didn't annotate the controller with @RestController but only with @Controller to make it perfectly clear, that this isn't a REST api, but a http Api, speaking JSON.

Also, one advantage here is the fact, that I can easily let the database do all the sorting and grouping. If I would like to return my own lists or maps of objects, I would have to take care that I would not mess the presorted collections up.