Create a simple UI that uses the Artist API
Go to the Oracle JET cookbook (http://www.oracle.com/webfolder/technetwork/jet/uiComponents-formControls.html) and read about the Select component. It can easily use our API from an AJAX call.

Could be swapped with an OJ ComboBox, so that the API can be used for searching items instead of doing this on the client