Add an example on how controllers directly depending on jOOQ can be tested
This is a nice example on why I prefer to abstract the DSLContext inside an @Repository or something similar. Such an @Repository could be mocked away with @MockBean and I could use @WebMvcTest for example.

It would be possible to use if I create different configuration classes for the test, but that would be a little too much now for the demo.